test(#614): repoint the refusal harness at a refusal it can still reach
`project.save_as` works now (#606), which broke the harness that watched it fail.
Driven at a writable path it succeeds, so there is no refusal to read: run that way it
failed 2 of 9 with `hint=''` and a project sitting at the path it asserts was never
written. A harness that cannot reach its subject is not evidence, and one that fails
for that reason trains everyone to ignore a red.

The refusal path did not go away; only this route to it did. So the run drives a path
under a directory that does NOT exist. `isValidProjectPath(requireExisting: false)`
lets it through, the menu opens the panel, and the panel's Go-to-Folder sheet will not
close on a path that does not resolve.

That target is strictly better than the one it replaces. The old refusal fired BEFORE
any panel opened, so "no panel is left behind" was green because nothing had ever been
there — the #606 harness says as much in its own comments. This one opens a panel and
clears it, so `panels_before_dismissal: 1 -> panels_after_dismissal: 0` is the
dismissal code running, observed, with a mutation that flips it.

Two harness defects fixed on the way, both of which made a check pass for the wrong
reason:

The "after" capture looked for the window by the title captured "before". A successful
save renames the document window, the lookup then finds nothing, and the visual
compares a real image against a failed lookup and calls it unchanged. Re-resolved by
frame area, the same fix the #606 harness got.

And the visual band is not what this directory has been calling it. `screencapture -l`
excludes the title bar, so the top 28px is window CONTENT: measured today, with the
Mixer open it lands on track names, the Inspector's Flex Mode row and a column of
plugin-slot buttons, and it carries a focus indicator that moves when a panel opens and
closes. A negative assertion over a region that changes for legitimate reasons produces
a red that means nothing. The band now has to prove it is quiet — two captures at rest
with nothing between them — before it is used as a negative control. If it is not, the
precondition and the visual both go red, which is the honest outcome: this proof cannot
be made from a window that was already moving. `expect_change` is NOT flipped to match
whatever happened; a check that agrees with its own result is not a check.

Renamed to live_614_the_refusal_it_can_still_reach.py, because the file no longer
watches what its old name claimed.
